Refactored parameter.
