+++ src/domain/models/parameters/Defaults.ts
@@ -1,0 +1,59 @@
+export const DEFAULTS = {
+    Ab: 'exec',
+    Pr: '1',
+    Jd: 'cond',
+    Tho: '0',
+    Abr: 'n',
+    Rec: '1',
+    Rei: '1',
+    Ha: 'n',
+    Eu: 'ent',
+    Soa: 'new',
+    Sea: 'new',
+    Mm: 'and',
+    Nmg: 'n',
+    Ega: 'none',
+    Uem: 'n',
+    Wc: 'no',
+    Shd: '2',
+    Sd: 'en',
+    Sdd: '1',
+    St: '+00:00',
+    Stt: '00:00',
+    Wt: 'no',
+    Cd: 'n',
+    Cftd: 'no',
+    Cgs: 'y',
+    Cond: 'and',
+    De: 'y',
+    Ej: 'gt',
+    Ejc: '0',
+    Eji: '0',
+    Ejm: 'ge',
+    Eju: 'le',
+    Etn: 'n',
+    Ets: 'kl',
+    Evesc: 'no',
+    Evspl: '10',
+    Evsrc: '0',
+    Evsrt: 'n',
+    Evssv: 'no',
+    Evtmc: 'n',
+    Flco: 'n',
+    Flwc: 'c',
+    Flwi: '60',
+    Fxg: 'none',
+    Gty: 'n',
+    Htexm: 'n',
+    Htknd: 'GET',
+    Htspt: 'diff',
+    Jty: 'q',
+    Mcs: 'm',
+    Md: 'th',
+    Mlsav: 'y',
+    Mp: 'n',
+    Ms: 'sch',
+    Ni: '-39',
+    Pfm: 'p',
+    Rg: '1',
+}
