+++ src/domain/models/units/Evwj.ts
@@ -22,7 +22,7 @@ export class Evwj extends UnitEntity {
     get evwfr() { return ParamFactory.evwfr(this); }
     // [evtmc={n|a|n:"file-name"|a:"file-name"
     //      |d:"file-name"|b:"file-name"};]
-    get evtmc() { return ParamFactory.evtmc(this, 'n'); }
+    get evtmc() { return ParamFactory.evtmc(this); }
     // [evuid=event-issue-source-user-ID;]
     get evuid() { return ParamFactory.evuid(this); }
     // [evgid=event-issue-source-group-ID;]
@@ -38,23 +38,23 @@ export class Evwj extends UnitEntity {
     // [ex="execution-agent-name";]
     get ex() { return ParamFactory.ex(this); }
     // [ha={y|n};]
-    get ha() { return ParamFactory.ha(this, 'n'); }
+    get ha() { return ParamFactory.ha(this); }
     // [eu={ent|def};]
-    get eu() { return ParamFactory.eu(this, 'ent'); }
+    get eu() { return ParamFactory.eu(this); }
     // [evesc={no|minutes};]
-    get evesc() { return ParamFactory.evesc(this, 'no'); }
+    get evesc() { return ParamFactory.evesc(this); }
     // [ets={kl|nr|wr|an};]
-    get ets() { return ParamFactory.ets(this, 'kl'); }
+    get ets() { return ParamFactory.ets(this); }
     // [mm={and|or};]
-    get mm() { return ParamFactory.mm(this, 'and'); }
+    get mm() { return ParamFactory.mm(this); }
     // [nmg={y|n};]
-    get nmg() { return ParamFactory.nmg(this, 'n'); }
+    get nmg() { return ParamFactory.nmg(this); }
     // [eun=name-of-the-unit-whose-end-is-being-waited-for;]
     get eun() { return ParamFactory.eun(this); }
     // [ega={exec|execdeffer|none};]
-    get ega() { return ParamFactory.ega(this, 'none'); }
+    get ega() { return ParamFactory.ega(this); }
     // [uem={y|n};]
-    get uem() { return ParamFactory.uem(this, 'n'); }
+    get uem() { return ParamFactory.uem(this); }
 
     /** Whether this jobnet have a unit whose end is being waited for. */
     get hasWaitedFor() {
