+++ src/domain/models/units/N.ts
@@ -23,9 +23,9 @@ export class N extends UnitEntity {
     }
 
     // [sd=[N,]{[[yyyy/]mm/]{[+|*|@]dd|[+|*|@]b[-DD]|[+]{su|mo|tu|we|th|fr|sa}[:{n|b}]}|en|ud};]
-    get sd() { return ParamFactory.sd(this, this.isRootJobnet ? 'en' : undefined); }
+    get sd() { return ParamFactory.sd(this); }
     // [st=[N,][+]hh:mm;]
-    get st() { return ParamFactory.st(this, '+00:00'); }
+    get st() { return ParamFactory.st(this); }
     // [sy=[N,]hh:mm|{M|U|C}mmmm;]
     get sy() { return ParamFactory.sy(this); }
     // [ey=[N,]hh:mm|{M|U|C}mmmm;]
@@ -37,33 +37,33 @@ export class N extends UnitEntity {
     // [sh=[N,]{be|af|ca|no};]
     get sh() { return ParamFactory.sh(this); }
     // [shd=[N,]n;]
-    get shd() { return ParamFactory.shd(this, '2'); }
+    get shd() { return ParamFactory.shd(this); }
     // [wt=[N,]{no|hh:mm|mmmm|un};]
-    get wt() { return ParamFactory.wt(this, 'no'); }
+    get wt() { return ParamFactory.wt(this); }
     // [wc=[N,]{no|n|un};]
-    get wc() { return ParamFactory.wc(this, 'no'); }
+    get wc() { return ParamFactory.wc(this); }
     // [cftd=[N,]{no|be|af|db|da}[,n[,N]];]
-    get cftd() { return ParamFactory.cftd(this, 'no'); }
+    get cftd() { return ParamFactory.cftd(this); }
     // [ed=yyyy/mm/dd;]
     get ed() { return ParamFactory.ed(this); }
     // [rg=n;]
-    get rg() { return ParamFactory.rg(this, this.isRootJobnet ? '1' : undefined); }
+    get rg() { return ParamFactory.rg(this); }
     // [pr=n;]
-    get pr() { return ParamFactory.pr(this, '1'); }
+    get pr() { return ParamFactory.pr(this); }
     // [ni=n;]
-    get ni() { return ParamFactory.ni(this, '-39'); }
+    get ni() { return ParamFactory.ni(this); }
     // [ha={y|w|a|n};]
-    get ha() { return ParamFactory.ha(this, 'n'); }
+    get ha() { return ParamFactory.ha(this); }
     // [ejn=exclusive-jobnet-name;]
     get ejn() { return ParamFactory.ejn(this); }
     // [cd={no|un|n};]
-    get cd() { return ParamFactory.cd(this, 'n'); }
+    get cd() { return ParamFactory.cd(this); }
     // [de={y|n};]
-    get de() { return ParamFactory.de(this, 'y'); }
+    get de() { return ParamFactory.de(this); }
     // [ms={sch|mlt};]
-    get ms() { return ParamFactory.ms(this, 'sch'); }
+    get ms() { return ParamFactory.ms(this); }
     // [mp={y|n};]
-    get mp() { return ParamFactory.mp(this, 'n'); }
+    get mp() { return ParamFactory.mp(this); }
     // [jc=full-path-job-group-name;]
     get jc() { return ParamFactory.jc(this); }
     // [rh="execution-manager-name";]
@@ -87,15 +87,15 @@ export class N extends UnitEntity {
     // [ncsv=connection-service-name;]
     get ncsv() { return ParamFactory.ncsv(this); }
     // [mm={and|or};]
-    get mm() { return ParamFactory.mm(this, 'and'); }
+    get mm() { return ParamFactory.mm(this); }
     // [nmg={y|n};]
-    get nmg() { return ParamFactory.nmg(this, 'n'); }
+    get nmg() { return ParamFactory.nmg(this); }
     // [eun=name-of-the-unit-whose-end-is-being-waited-for;]
     get eun() { return ParamFactory.eun(this); }
     // [ega={exec|execdeffer|none};]
-    get ega() { return ParamFactory.ega(this, 'none'); }
+    get ega() { return ParamFactory.ega(this); }
     // [uem={y|n};]
-    get uem() { return ParamFactory.uem(this, 'n'); }
+    get uem() { return ParamFactory.uem(this); }
 
     get priority(): number {
         const pr = this.pr;
