+++ src/domain/models/units/Tmwj.ts
@@ -5,7 +5,7 @@ export class Tmwj extends UnitEntity {
     // [tmitv=wait-time;] 
     get tmitv() { return ParamFactory.tmitv(this); }
     // [etn={y|n};] 
-    get etn() { return ParamFactory.etn(this, 'n'); }
+    get etn() { return ParamFactory.etn(this); }
     // [jpoif=macro-variable-name:passing-information-name;] 
     get jpoif() { return ParamFactory.jpoif(this); }
     // [etm=n;] 
@@ -15,21 +15,21 @@ export class Tmwj extends UnitEntity {
     // [ex="execution-agent-name";] 
     get ex() { return ParamFactory.ex(this); }
     // [ha={y|n};] 
-    get ha() { return ParamFactory.ha(this, 'n'); }
+    get ha() { return ParamFactory.ha(this); }
     // [eu={ent|def};] 
-    get eu() { return ParamFactory.eu(this, 'ent'); }
+    get eu() { return ParamFactory.eu(this); }
     // [ets={kl|nr|wr|an};] 
-    get ets() { return ParamFactory.ets(this, 'kl'); }
+    get ets() { return ParamFactory.ets(this); }
     // [mm={and|or};] 
-    get mm() { return ParamFactory.mm(this, 'and'); }
+    get mm() { return ParamFactory.mm(this); }
     // [nmg={y|n};] 
-    get nmg() { return ParamFactory.nmg(this, 'n'); }
+    get nmg() { return ParamFactory.nmg(this); }
     // [eun=name-of-the-unit-whose-end-is-being-waited-for;] 
     get eun() { return ParamFactory.eun(this); }
     // [ega={exec|execdeffer|none};] 
-    get ega() { return ParamFactory.ega(this, 'none'); }
+    get ega() { return ParamFactory.ega(this); }
     // [uem={y|n};] 
-    get uem() { return ParamFactory.uem(this, 'n'); }
+    get uem() { return ParamFactory.uem(this); }
 
     /** Whether this jobnet have a unit whose end is being waited for. */
     get hasWaitedFor() {
