+++ src/ui-component/editor/ajsTable/columnDefs/group5.tsx
@@ -3,8 +3,8 @@ import { ColumnHelper, GroupColumnDef } from "@tanstack/table-core";
 import * as ajscolumn from '@resource/i18n/ajscolumn';
 import { UnitEntity } from "../../../../domain/models/units/UnitEntities";
 import { defaultAccessorFn, tyAccessorFn } from "./common";
-import { Gty } from "../../../../domain/models/parameters/ParameterEntities";
 import { Chip } from "@mui/material";
+import { Gty } from "../../../../domain/models/parameters";
 
 const group5 = (
     columnHelper: ColumnHelper<UnitEntity>,
